+++ sample/particlesHDR/meta.ts
