+++ sample/particlesHDR/particle.wgsl
