+++ sample/particlesHDR/probabilityMap.wgsl
